Replace files and make dataset abstract class
